% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdc-package.R
\docType{package}
\name{acdc-package}
\alias{acdc}
\alias{acdc-package}
\title{acdc package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This R package provides processes for supporting data calls productions. These processes go through extractions and links from Ob7 databases, quality controls, descriptive analysis and data design.
}
\section{FDI processes and functions}{

\itemize{
\item{\code{\link{fdi_tablea_catch_summary}}}
\item{\code{\link{fdi_tabled_discard_length}}}
\item{\code{\link{fdi_tablef_landings_length}}}
\item{\code{\link{fdi_tableg_effort}}}
\item{\code{\link{fdi_tableh_landings_rectangle}}}
\item{\code{\link{fdi_tablei_effort_rectangle}}}
\item{\code{\link{fdi_tablej_capacity}}}
\item{\code{\link{fdi_shortcut_function}}}
\item{\code{\link{fdi_template_checking}}}
\item{\code{\link{fdi_tables_consistency}}}
\item{\code{\link{fdi_table_export}}}
\item{\code{\link{fdi_template_identification}}}
}
}

\section{ICCAT processes and functions}{

\itemize{
\item{\code{\link{iccat_st01_t1fc}}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/acdc}
  \item \url{https://ob7-ird.github.io/acdc/}
  \item Report bugs at \url{https://github.com/OB7-IRD/acdc/issues}
}

}
\author{
\strong{Maintainer}: Mathieu Depetris \email{mathieu.depetris@ird.fr} (\href{https://orcid.org/0000-0001-8080-0531}{ORCID})

Authors:
\itemize{
  \item Laurent Floc'h \email{laurent.floch@ird.fr} (\href{https://orcid.org/0000-0003-4764-9766}{ORCID})
  \item Philippe Sabarros \email{philippe.sabarros@ird.fr} (\href{https://orcid.org/0000-0001-6135-9015}{ORCID})
}

Other contributors:
\itemize{
  \item Julien Lebranchu \email{julien.lebranchu@ird.fr} (\href{https://orcid.org/0000-0002-7759-0534}{ORCID}) [contributor]
  \item IRD - Ob7  Observatory of Exploited Tropical Pelagic Ecosystems \email{ob7@ird.fr} (\href{https://orcid.org/0000-0002-4642-3418}{ORCID}) [copyright holder, funder]
}

}
\keyword{internal}
