% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments_controls.R
\name{fdi_r_type_checking}
\alias{fdi_r_type_checking}
\alias{fdi_table_id_checking}
\title{FDI R type checking}
\usage{
fdi_table_id_checking(table_id)
}
\description{
FDI R type checking
}
\keyword{internal}
