% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdi_shortcut_function.R
\name{fdi_shortcut_function}
\alias{fdi_shortcut_function}
\title{FDI tables shortcut function (FDI process)}
\usage{
fdi_shortcut_function(
  balbaya_con,
  sardara_con,
  t3_con,
  period,
  gear,
  flag,
  observe_bycatch_path,
  observe_discard_path,
  tablea_catch_summary,
  tablea_bycatch_retained,
  tablea_landing_rectangle,
  tableg_effort_rectangle,
  fao_area_file_path,
  eez_area_file_path,
  cwp_grid_1deg_1deg,
  cwp_grid_5deg_5deg,
  template_checking = FALSE,
  template_year = NULL,
  table_export_path = NULL
)
}
\arguments{
\item{balbaya_con}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} for a connection to the balbaya database.}

\item{sardara_con}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} for a connection to the sardara database.}

\item{t3_con}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} for a connection to the t3 database.}

\item{period}{{\link[base]{integer}} expected. Year period for data extractions.}

\item{gear}{{\link[base]{integer}}. Gear(s) selection for data extractions.}

\item{flag}{{\link[base]{integer}} expected. Flag(s) selection for data extractions.}

\item{observe_bycatch_path}{{\link[base]{character}} expected. Directory path of the bycatch data extractions. Check this input with Philippe Sabarros (philippe.sabarros@ird.fr).}

\item{observe_discard_path}{{\link[base]{character}} expected. Directory path of the discards data extractions. Check this input with Philippe Sabarros (philippe.sabarros@ird.fr).}

\item{tablea_catch_summary}{{\link[base]{list}} expected. Output "table_a" of the function {\link[acdc]{fdi_tablea_catch_summary}}.}

\item{tablea_bycatch_retained}{Output "bycatch_retained" of the function {\link[acdc]{fdi_tablea_catch_summary}}.}

\item{tablea_landing_rectangle}{Output "landing_rectangle" of the function {\link[acdc]{fdi_tablea_catch_summary}}.}

\item{tableg_effort_rectangle}{Output "effort_rectangle" of the function {\link[acdc]{fdi_tableg_effort}}.}

\item{fao_area_file_path}{{\link[base]{character}} expected. File path of the FAO area grid. The file format has to be .Rdata.}

\item{eez_area_file_path}{{\link[base]{character}} expected. File path of the EEZ area grid. The file format has to be .Rdata.}

\item{cwp_grid_1deg_1deg}{{\link[base]{character}} expected. File path of the CWP area grid of 1째 by 1째. The file format has to be .Rdata.}

\item{cwp_grid_5deg_5deg}{{\link[base]{character}} expected. File path of the CWP area grid of 5째 by 5째. The file format has to be .Rdata.}

\item{template_checking}{{\link[base]{logical}} expected. By default FALSE Checking FDI table generated regarding the official FDI template.}

\item{template_year}{{\link[base]{integer}} expected. By default NULL. Template year.}

\item{table_export_path}{{\link[base]{character}} expected. By default NULL. Directory path associated for the export.}
}
\description{
Shortcut function to generate all FDI tables.
}
