% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdi_table_export.R
\name{fdi_table_export}
\alias{fdi_table_export}
\title{Table export (FDI process)}
\usage{
fdi_table_export(fdi_table, export_path, table_id)
}
\arguments{
\item{fdi_table}{R object expected. FDI table selected for export process.}

\item{export_path}{{\link[base]{character}} expected. Directory path associated for the export.}

\item{table_id}{{\link[base]{character}} expected. Identification of the FDI table.}
}
\description{
Process for export in xlsx format FDI tables.
}
