% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdi_tablea_catch_summary.R
\name{fdi_tablea_catch_summary}
\alias{fdi_tablea_catch_summary}
\title{Number of vessels}
\usage{
fdi_tablea_catch_summary(
  balbaya_con,
  observe_bycatch_path,
  periode,
  gear,
  flag,
  fao_area_file_path,
  eez_area_file_path,
  cwp_grid_file_path,
  template_checking = TRUE,
  template_year = NULL,
  table_export_path = NULL
)
}
\arguments{
\item{dbconnection}{{\link[base]{list}} or {\link[base]{character}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} or name of a internal fishi dataset.}

\item{countries}{{\link[base]{integer}} expected. Countries codes identifiaction.}

\item{oceans}{{\link[base]{integer}} expected. Oceans codes identifiaction.}

\item{vessel_types}{{\link[base]{integer}}. Vessel types codes identification.}

\item{period}{{\link[base]{integer}} expected. Period identification in year.}

\item{time_step}{{\link[base]{character}} expected. Kind of display you want in the graphic output. You can choose between "month" and "year".}
}
\description{
Number of vessels by country, ocean, period, time step and vessel type.
}
