% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdi_tabled_discard_length.R
\name{fdi_tabled_discard_length}
\alias{fdi_tabled_discard_length}
\title{Table D NAO OFR discards length generation (FDI process)}
\usage{
fdi_tabled_discard_length(
  observe_discard_path,
  tablea_catch_summary,
  template_checking = TRUE,
  template_year = NULL,
  table_export_path = NULL
)
}
\arguments{
\item{observe_discard_path}{{\link[base]{character}} expected. Directory path of the discards data extractions. Check this input with Philippe Sabarros (philippe.sabarros@ird.fr).}

\item{tablea_catch_summary}{{\link[base]{list}} expected. Output "table_a" of the function {\link[acdc]{fdi_tablea_catch_summary}}.}

\item{template_checking}{{\link[base]{logical}} expected. By default TRUE. Checking FDI table generated regarding the official FDI template.}

\item{template_year}{{\link[base]{integer}} expected. By default NULL. Template year.}

\item{table_export_path}{{\link[base]{character}} expected. By default NULL. Directory path associated for the export.}
}
\value{
The process returns a list with the FDI table D inside.
}
\description{
Process for generation and optionally extraction of the FDI table D (NAO OFR discards length).
}
