% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdi_tablef_landings_length.R
\name{fdi_tablef_landings_length}
\alias{fdi_tablef_landings_length}
\title{Table F NAO OFR landings length generation (FDI process)}
\usage{
fdi_tablef_landings_length(
  balbaya_con,
  sardara_con,
  t3_con,
  period,
  gear,
  flag,
  tablea_bycatch_retained,
  tablea_catch_summary,
  cwp_grid_file_path,
  fao_area_file_path,
  template_checking = FALSE,
  template_year = NULL,
  table_export_path = NULL
)
}
\arguments{
\item{balbaya_con}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} for a connection to the balbaya database.}

\item{sardara_con}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} for a connection to the sardara database.}

\item{t3_con}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} for a connection to the t3 database.}

\item{period}{{\link[base]{integer}} expected. Year period for data extractions.}

\item{gear}{{\link[base]{integer}}. Gear(s) selection for data extractions.}

\item{flag}{{\link[base]{integer}} expected. Flag(s) selection for data extractions.}

\item{tablea_bycatch_retained}{Output "bycatch_retained" of the function {\link[acdc]{fdi_tablea_catch_summary}}.}

\item{tablea_catch_summary}{Output "table_a" of the function {\link[acdc]{fdi_tablea_catch_summary}}.}

\item{cwp_grid_file_path}{{\link[base]{character}} expected. File path of the CWP area grid. The file format has to be .Rdata or .RData extension.}

\item{fao_area_file_path}{{\link[base]{character}} expected. File path of the FAO area grid. The file format has to be .Rdata or RData extension.}

\item{template_checking}{{\link[base]{logical}} expected. By default FALSE Checking FDI table generated regarding the official FDI template.}

\item{template_year}{{\link[base]{integer}} expected. By default NULL. Template year.}

\item{table_export_path}{{\link[base]{character}} expected. By default NULL. Directory path associated for the export.}
}
\value{
The process returns a list with the FDI table F inside.
}
\description{
Process for generation and optionally extraction of the FDI table F (NAO OFR landings length).
}
