% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdi_tablei_effort_rectangle.R
\name{fdi_tablei_effort_rectangle}
\alias{fdi_tablei_effort_rectangle}
\title{Table I effort by rectangle generation (FDI process)}
\usage{
fdi_tablei_effort_rectangle(
  tableg_effort_rectangle,
  template_checking = TRUE,
  template_year = NULL,
  table_export_path = NULL
)
}
\arguments{
\item{tableg_effort_rectangle}{Output "effort_rectangle" of the function {\link[acdc]{fdi_tableg_effort}}.}

\item{template_checking}{{\link[base]{logical}} expected. By default TRUE. Checking FDI table generated regarding the official FDI template.}

\item{template_year}{{\link[base]{integer}} expected. By default NULL. Template year.}

\item{table_export_path}{{\link[base]{character}} expected. By default NULL. Directory path associated for the export.}
}
\value{
The process returns a list with the FDI table I inside.
}
\description{
Process for generation and optionally extraction of the FDI table I (effort by rectangle).
}
