% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdi_tables_consistency.R
\name{fdi_tables_consistency}
\alias{fdi_tables_consistency}
\title{FDI tables consistency (FDI process)}
\usage{
fdi_tables_consistency(
  tablea,
  tabled,
  tablef,
  tableg,
  tableh,
  tablei,
  tablea_bycatch_retained
)
}
\arguments{
\item{tablea}{Output "table_a" of the function {\link[acdc]{fdi_tablea_catch_summary}}}

\item{tabled}{Output "table_d" of the function {\link[acdc]{fdi_tabled_discard_length}}}

\item{tablef}{Output "table_f" of the function {\link[acdc]{fdi_tablef_landings_length}}}

\item{tableg}{Output "table_g" of the function {\link[acdc]{fdi_tableg_effort}}}

\item{tableh}{Output "table_h" of the function {\link[acdc]{fdi_tableh_landings_rectangle}}}

\item{tablei}{Output "table_i" of the function {\link[acdc]{fdi_tablei_effort_rectangle}}}

\item{tablea_bycatch_retained}{Output "bycatch_retained" of the function {\link[acdc]{fdi_tablea_catch_summary}}.}
}
\value{
The process returns a list with 3 verifications data.frame
}
\description{
Process for cross-checking of the FDI tables.
}
