% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdi_template_checking.R
\name{fdi_template_checking}
\alias{fdi_template_checking}
\title{FDI template checking (FDI process)}
\usage{
fdi_template_checking(fdi_table, template_year, table_id)
}
\arguments{
\item{fdi_table}{Output table from a FDI process.}

\item{template_year}{{\link[base]{integer}} expected. Template year.}

\item{table_id}{{\link[base]{character}} expected. Identification of the FDI table.}
}
\description{
Process for template checking of a FDI table.
}
