% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdi_template_identification.R
\name{fdi_template_identification}
\alias{fdi_template_identification}
\title{Template identification (FDI process)}
\usage{
fdi_template_identification(table_id)
}
\arguments{
\item{table_id}{{\link[base]{character}} expected. Identification of the FDI table.}
}
\description{
Process for template identification of FDI table.
}
