% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments_controls.R
\name{global_database_connection_checking}
\alias{global_database_connection_checking}
\title{Global database connection checking}
\usage{
global_database_connection_checking(database_con)
}
\description{
Global database connection checking
}
\keyword{internal}
