% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments_controls.R
\name{global_export_path_checking}
\alias{global_export_path_checking}
\title{Global export path checking}
\usage{
global_export_path_checking(export_path)
}
\description{
Global export path checking
}
\keyword{internal}
