% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments_controls.R
\name{global_integer_checking}
\alias{global_integer_checking}
\title{Global integer checking}
\usage{
global_integer_checking(integer_object)
}
\description{
Global integer checking
}
\keyword{internal}
