% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments_controls.R
\name{global_logical_checking}
\alias{global_logical_checking}
\title{Global logical checking}
\usage{
global_logical_checking(logical_object)
}
\description{
Global logical checking
}
\keyword{internal}
