% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments_controls.R
\name{global_rdata_type_checking}
\alias{global_rdata_type_checking}
\title{Global RData type checking}
\usage{
global_rdata_type_checking(rdata_path)
}
\description{
Global RData type checking
}
\keyword{internal}
