% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridSearch_pcs_fast.R
\name{GridSearch_pcs_fast}
\alias{GridSearch_pcs_fast}
\title{Determination of the optimal combination of k and resolution for
single-cell clustering.}
\usage{
GridSearch_pcs_fast(
  object,
  assay.name,
  .resolutions = seq(0.1, 1.9, by = 0.2),
  .bootstraps = 1,
  .knns = seq(11, 101, by = 10),
  .pct_cells = 100,
  .replace = FALSE,
  .clust_alg = "Louvain",
  free.cores = 2,
  .type = "genes",
  .dims = NULL,
  my_seed = 0,
  weights = "unitary",
  exp_base = 2.718282,
  show_progress_bar = TRUE
)
}
\arguments{
\item{object}{Seurat object (genes/proteins along rows and cells on columns)}

\item{assay.name}{Assay to use}

\item{.resolutions}{vector of the resolution parameter. Default=seq(0.01,2,by = 0.01).}

\item{.bootstraps}{vector of bootstraps. Default=1 (i.e. no bootstraps).}

\item{.knns}{sequence of values for the number of nearest neighbors. Default=seq(3, 31, by=2).}

\item{.pct_cells}{percentage of cells sample at each bootstrap iteration. All cells are used by default.}

\item{.clust_alg}{clustering algorithm. Choose among: "Louvain" (default); "Louvain-mult-ref"; "SLM"; "Leiden".}

\item{my_seed}{random seed for FindClusters. Default=0.}

\item{weights}{unitary (unitary) or exponential (exp) way of weighing the silhouette scores. Default=unitary.}

\item{.replace:}{(logical) whether to sample cells with (TRUE) or without replacement (FALSE). Default=FALSE.}

\item{free_cores}{number of cores that are not used for the calculation. Default=2.}
}
\value{
A 11-columns tibble containing the outcomes of the calculation for choosing the optimal number of clusters.
Columns consist of the following vectors.
\itemize{
\item `index`: integer in the interval 1-`n_idx`, where `n_idx`=`length(.bootstrap)*length(.knns)`
uniquely assigned to each combination of `.bootstraps` and `.knns`.
\item `boostrap`: integer assigned to the given resampling. Possible values are those in `.bootstraps`.
\item `knn`: number of nearest neighbors used as input parameters to `FindNeighbors`. Possible values are those in `.knns`.
\item `resolution`: resolution for `FindClusters`. Possible values are those in `.resolutions`.
\item `tot_sil_neg`: negative silhouette score
\item `lowest_sil_clust`: lowest silhouette
\item `sil_avg`: average silhouette score
\item `sil_mean_median`: mean of the median silhouette score
\item `n_clust`: number of clusters resulting for the given combination of `knn`, `resolution` with cells sampled according to `bootstrap`.
\item `random.seed`: random seed used to initialize the rng for cell subsamplings. Values are the same as `index`.
}
}
\description{
`GridSearch_pcs_fast()` optimizes clustering solution of scRNA-seq databy
grid-search optimization of the Silhouette score across multiple
resamplings. Runs faster by computing the distance matrix once if not bootstrapping
and placing NN in the outer loop and resolution in the inner loop, reducing the number
of times that neighbors are computed.
}
\details{
Input arguments
}
