% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSCA_par.R
\name{OSCA_par}
\alias{OSCA_par}
\title{OSCA_par}
\usage{
OSCA_par(
  df,
  externalVar,
  ILCincrement = 0.05,
  oscaPath,
  numNodes = 1,
  permute = T
)
}
\arguments{
\item{df}{n x p data frame or matrix of gene expression values with no ID column}

\item{externalVar}{vector of length n of external variable values with no ID column}

\item{ILCincrement}{float between zero and one determining interval between tested ILC values; default is 0.05}

\item{oscaPath}{absolute path to OSCA software}

\item{numNodes}{number of available compute nodes for parallelization; default is 1}

\item{permute}{boolean value for whether or not to include permutations; default is true}
}
\value{
Data frame with columns

\describe{
\item{ILC}{the information loss criterion used for that iteration}
\item{InformationLost}{percent information lost due to data reduction}
\item{PercentReduction}{percent of variables condensed compared to unreduced data}
\item{VarianceExplained_Observed}{percent variance explained in observed external variable by the data}
\item{SE_Observed}{standard error of the percent variance estimate for observed external variable}
\item{VarianceExplained_Permuted}{percent variance explained in permuted external variable by the data; only if input parameter "permute" is true}
\item{SE_Permuted}{standard error of the percent variance estimate for permuted external variable; only if input parameter "permute" is true}
}
}
\description{
Function to return a dataframe of ILCs and percent variance explained in an external phenotype for a range of ILCs between zero and one.
}
\examples{
#load CCA package for example dataset
library(CCA)

# load dataset
data("nutrimouse")

# generate random phenotype
r.pheno <- rnorm(nrow(nutrimouse$gene))

## random phenotype
# run function; input path to OSCA software
# OSCA_par(df = nutrimouse$gene, 
#          externalVar = r.pheno, 
#          ILCincrement = 0.25, 
#          oscaPath = "pathHere", 
#          numNodes = detectCores()-1, 
#          permute = T)

## observed external variable
# run function; input path to OSCA software
# OSCA_par(df = nutrimouse$gene, 
#          externalVar = as.numeric(nutrimouse$diet),
#          ILCincrement = 0.25, 
#          oscaPath = "pathHere", 
#          numNodes = detectCores()-1, 
#          permute = T)

}
\references{
Benjamini Y, Hochberg Y. Controlling the false discovery rate: a practical and powerful approach to multiple testing. Journal of the Royal statistical society: series B (Methodological) 57 (1995) 289–300.

Martin P, et al. Novel aspects of PPARalpha-mediated regulation of lipid and xenobiotic metabolism revealed through a nutrigenomic study. Hepatology, in press, 2007.

Millstein J, Battaglin F, Barrett M, Cao S, Zhang W, Stintzing S, et al. Partition: a surjective mapping approach for dimensionality reduction. Bioinformatics 36 (2019) 676–681. doi:10.1093/bioinformatics/ btz661.
}
\seealso{
OSCA software - \url{https://yanglab.westlake.edu.cn/software/osca}
}
\author{
Katelyn Queen, \email{kjqueen@usc.edu}
}
