% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSCA_parPlot.R
\name{OSCA_parPlot}
\alias{OSCA_parPlot}
\title{OSCA_parPlot}
\usage{
OSCA_parPlot(df, externalVarName = "", dataName = "")
}
\arguments{
\item{df}{output from OSCA_par function with permutations}

\item{externalVarName}{string of name of external variable for graph labels; default is blank}

\item{dataName}{string of name of data for graph labels; default is blank}
}
\value{
ggplot object
}
\description{
Function to return a graph comparing percent variance explained in an external phenotype and information lost/percent reduction for both observed and permuted data
}
\references{
\itemize{
\item BH FDR: \url{https://www.jstor.org/stable/2346101}
\item Partition: \url{https://pubmed.ncbi.nlm.nih.gov/31504178/}
}
}
\seealso{
OSCA software - \url{https://yanglab.westlake.edu.cn/software/osca}
}
\author{
Katelyn Queen, \email{kjqueen@usc.edu}
}
