% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbes_ce.R
\name{rdbes_ce}
\alias{rdbes_ce}
\title{Table Commercial Effort (CE) generation (RDBES process)}
\usage{
rdbes_ce(
  observe_con,
  balbaya_con,
  fao_area_file_path,
  eez_area_file_path,
  year_time_period,
  flag,
  major_fao_area_filter = NULL,
  hash_algorithms = NULL,
  export_path = NULL
)
}
\arguments{
\item{observe_con}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} for a connection to the observe database.}

\item{balbaya_con}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} for a connection to the balbaya database.}

\item{fao_area_file_path}{{\link[base]{character}} expected. File path of the FAO area grid. The file format has to be .Rdata or .RData extension.}

\item{eez_area_file_path}{{\link[base]{character}} expected. File path of the EEZ area grid. The file format has to be .Rdata or .RData extension.}

\item{year_time_period}{{\link[base]{integer}} expected. Year(s) selected associated to the databases queries extractions.}

\item{flag}{{\link[base]{integer}} expected. Flag(s) selected associated to the databases queries extractions.}

\item{major_fao_area_filter}{{\link[base]{integer}} expected. By default NULL. Sub selection of major fao area.}

\item{hash_algorithms}{{\link[base]{integer}} expected. By default NULL. The hashing algorithms to be used for the CEencrypVesIds variable. You can choose any modality of the argument "algo" or the function {\link[digest]{digest}}.}

\item{export_path}{{\link[base]{character}} expected. By default NULL. Directory path associated for the export.}
}
\value{
A R object with the RDBES table CE with potentially a csv extraction associated.
}
\description{
Process for generation and optionally extraction of the RDBES table CE (Commercial Effort).
}
