% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbes_cl.R
\name{rdbes_cl}
\alias{rdbes_cl}
\title{Table Commercial Landing (CL) generation (RDBES process)}
\usage{
rdbes_cl(
  observe_con,
  balbaya_con,
  fao_area_file_path,
  eez_area_file_path,
  year_time_period,
  flag,
  export_path = NULL
)
}
\arguments{
\item{observe_con}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} for a connection to the observe database.}

\item{balbaya_con}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} for a connection to the balbaya database.}

\item{fao_area_file_path}{{\link[base]{character}} expected. File path of the FAO area grid. The file format has to be .Rdata or .RData extension.}

\item{eez_area_file_path}{{\link[base]{character}} expected. File path of the EEZ area grid. The file format has to be .Rdata or .RData extension.}

\item{year_time_period}{{\link[base]{integer}} expected. Year(s) selected associated to the databases queries extractions.}

\item{flag}{{\link[base]{integer}} expected. Flag(s) selected associated to the databases queries extractions.}

\item{export_path}{{\link[base]{character}} expected. By default NULL. Directory path associated for the export.}
}
\value{
A R object with the RDBES table CL with potentially a csv extraction associated.
}
\description{
Process for generation and optionally extraction of the RDBES table CL (Commercial Landing).
}
