% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbes_vd.R
\name{rdbes_vd}
\alias{rdbes_vd}
\title{Table Vessel Details (VD) generation (RDBES process)}
\usage{
rdbes_vd(observe_con, rdbes_cl, flag, export_path = NULL)
}
\arguments{
\item{observe_con}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} for a connection to the observe database.}

\item{flag}{{\link[base]{integer}} expected. Flag(s) selected associated to the databases queries extractions.}

\item{export_path}{{\link[base]{character}} expected. By default NULL. Directory path associated for the export.}
}
\value{
A R object with the RDBES table VD with potentially a csv extraction associated.
}
\description{
Process for generation and optionally extraction of the RDBES table VD (Vessel Details).
}
