% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwp_table_2_1.R
\name{rwp_table_2_1}
\alias{rwp_table_2_1}
\title{Regional Work Plan table 2.1}
\usage{
rwp_table_2_1(
  reference_period_start,
  reference_period_end,
  country,
  eurostat,
  rcg_stats,
  national_stats,
  fides,
  fides_common,
  rfmo,
  output_suffix = NULL,
  path_input,
  path_in_confidential_data,
  path_output
)
}
\arguments{
\item{reference_period_start}{{\link[base]{integer}} expected. Start of reference period. Be careful, the process needs 3 years at least to run.}

\item{reference_period_end}{{\link[base]{integer}} expected. End of reference period. Be careful, the process needs 3 years at least to run.}

\item{country}{{\link[base]{character}} expected. Country of interest. Use 2-alpha country.}

\item{eurostat}{{\link[base]{logical}} expected. Landing statistics downloaded from EUROSTAT (https://ec.europa.eu/eurostat).}

\item{rcg_stats}{{\link[base]{logical}} expected. Landing statistics from the regional database.}

\item{national_stats}{{\link[base]{logical}} expected. National landing statistics.}

\item{fides}{{\link[base]{logical}} expected. Use national extraction from FIDES including a country and EEC.}

\item{fides_common}{{\link[base]{logical}} expected. FIDEScommon is a Regional Coordination Group (RCG) extraction from FIDES and including all EU countries}

\item{rfmo}{{\link[base]{character}} expected. RFMO's list to include in output.}

\item{output_suffix}{{\link[base]{character}} expected. Add an suffic to the output name.}

\item{path_input}{{\link[base]{character}} expected. Input path for input files.}

\item{path_in_confidential_data}{{\link[base]{character}} expected. Input path for confidential files.}

\item{path_output}{{\link[base]{character}} expected. Output path for output files.}
}
\value{
I don't know yet dude!
}
\description{
Process for generation the table 2.1, list of required species/stocks, for a Regional Work Plan (RWP).
}
