% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_tools.R
\name{SAClustering}
\alias{SAClustering}
\title{SAClustering
Computes silhouette-optimized clustering solution using Simulated Annealing.}
\usage{
SAClustering(
  S.obj,
  NN_range = c(3, 30),
  assay = "RNA",
  slot = "scale.data",
  clust_alg = 1,
  type.fun = "mean.silhouette",
  reduction = FALSE,
  control = NULL,
  verbose = TRUE,
  final = TRUE,
  plot = TRUE
)
}
\arguments{
\item{S.obj}{a Seurat object}

\item{NN_range}{numeric vector containing the minimum and maximum number of nearest neighbors to
on which optimization should be performed.}

\item{assay}{assay}
}
\description{
'SA_tools' returns
}
\details{
I hope it is gonna work
}
