% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.R
\name{ACDC}
\alias{ACDC}
\title{ACDC}
\usage{
ACDC(fullData, ILC = 0.5, externalVar, identifierList)
}
\arguments{
\item{fullData}{data frame or matrix with samples as rows, all features as columns}

\item{ILC}{information loss criterion for Partition, or the minimum intra-class correlation required for features to be condensed; 0 <= ILC <= 1; default is 0.50}

\item{externalVar}{data frame, matrix, or vector containing external variable data to be used for CCA, rows are samples}

\item{identifierList}{row vector of identifiers corresponding to columns in fullData (ex: Hugo symbols for genes); optional}
}
\value{
Data frame, sorted by ascending BH FDR value, with columns

\describe{
\item{moduleNum}{module number}
\item{features}{list of identifiers from input parameter "identifierList" for all features in the module; 0 if identifierList not included}
\item{CCA_corr}{list of CCA canonical correlation coefficients}
\item{CCA_pval}{Wilks-Lamda F-test p-value or t-test p-value (see details)}
\item{BHFDR_qval}{Benjamini-Hochberg false discovery rate q-value}
}
}
\description{
Function for association of covariance to detect differential co-expression analysis of modules of correlated genes or molecular features and external data with modules to be defined by Partition.
}
\references{
\itemize{
\item BH FDR: \url{https://www.jstor.org/stable/2346101}
\item Partition: \url{https://pubmed.ncbi.nlm.nih.gov/31504178/}
}
}
\author{
Katelyn Queen, \email{kjqueen@usc.edu}
}
