% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.R
\name{ACDC}
\alias{ACDC}
\title{ACDC}
\usage{
ACDC(fullData, ILC = 0.5, externalVar, geneList)
}
\arguments{
\item{fullData}{data frame or matrix with samples as rows, all probes as columns}

\item{ILC}{information loss criterion for Partition; 0 <= ILC <= 1; default is 0.50}

\item{externalVar}{data frame or matrix containing external variable data to be used for CCA, rows are samples}

\item{geneList}{row vector of gene names corresponding to columns in fullData}
}
\value{
Data frame, sorted by ascending BH FDR value, with columns

\describe{
\item{moduleNum}{module number}
\item{genes}{list of Hugo gene symbols for all genes in the module}
\item{CCA_corr}{list of CCA canonical correlation coefficients}
\item{CCA_pval}{Wilks-Lamda F-test p-value}
\item{BHFDR_qval}{Benjamini-Hochberg false discovery rate q-value}
}
}
\description{
Function for association of covariance to detect differential co-expression analysis of modules of correlated genes and external data using Partition
}
\references{
\itemize{
\item BH FDR: \url{https://www.jstor.org/stable/2346101}
\item Partition: \url{https://pubmed.ncbi.nlm.nih.gov/31504178/}
}
}
\author{
Katelyn Queen, \email{kjqueen@usc.edu}
}
