% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coVar.R
\name{coVar}
\alias{coVar}
\title{coVar}
\usage{
coVar(dataPair, fullData)
}
\arguments{
\item{dataPair}{column indices of two genes to calculate covariance between}

\item{fullData}{dataframe or matrix with samples as rows, all probes as columns}
}
\value{
Vector of pairwise covariances for all samples
}
\description{
Function to calculate ACDC covariances within a data pair for all samples
}
\author{
Katelyn Queen, \email{kjqueen@usc.edu}
}
