% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdi_tabled_discard_length.R
\name{fdi_tabled_discard_length}
\alias{fdi_tabled_discard_length}
\title{Table D NAO OFR discards length generation (FDI process)}
\usage{
fdi_tabled_discard_length(
  observe_con,
  period,
  ocean,
  program,
  flag,
  fao_area_file_path,
  tablea_catch_summary,
  template_checking = FALSE,
  template_year = NULL,
  table_export_path = NULL
)
}
\arguments{
\item{observe_con}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} for a connection to the observe database.}

\item{period}{{\link[base]{integer}} expected. Year(s) period for data extraction(s).}

\item{ocean}{{\link[base]{character}} expected. Ocean(s) of interest.}

\item{program}{{\link[base]{character}} expected. Programs of interest.}

\item{flag}{{\link[base]{integer}} expected. Flag(s) selection for data extraction(s).}

\item{fao_area_file_path}{{\link[base]{character}} expected. File path of the FAO area grid. The file format has to be .Rdata or .RData extension.}

\item{tablea_catch_summary}{{\link[base]{list}} expected. Output "table_a" of the function {\link[acdc]{fdi_tablea_catch_summary}}.}

\item{template_checking}{{\link[base]{logical}} expected. By default FALSE Checking FDI table generated regarding the official FDI template.}

\item{template_year}{{\link[base]{integer}} expected. By default NULL. Template year.}

\item{table_export_path}{{\link[base]{character}} expected. By default NULL. Directory path associated for the export.}
}
\value{
The process returns a list with the FDI table D inside.
}
\description{
Process for generation and optionally extraction of the FDI table D (NAO OFR discards length).
}
