% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_load_fides_data.R
\name{global_load_fides_data}
\alias{global_load_fides_data}
\title{Load FIDES data}
\usage{
global_load_fides_data(
  fides_format = TRUE,
  reference_period,
  path = NULL,
  country = NULL
)
}
\arguments{
\item{fides_format}{{\link[base]{logical}} expected. Specify her if you want to load data from "official" FIDES files.}

\item{reference_period}{{\link[base]{integer}} expected. Period of reference, in years.}

\item{path}{{\link[base]{character}} expected. Input path directory where FIDES files are located.}

\item{country}{{\link[base]{character}} expected. Country(ies) id(s) for data extraction associated. Use 3-alpha country.}
}
\value{
The function return a {\link[tibble]{tibble}}.
}
\description{
Process for Load FIDES data from "official" FIDES files or a public alternative available on https://griffincarpenter.org/reports/european-fishing-quotas-2001-2021/.
}
