% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSCA_parPlot.R
\name{OSCA_parPlot}
\alias{OSCA_parPlot}
\title{OSCA_parPlot}
\usage{
OSCA_parPlot(df, externalVarName = "", dataName = "")
}
\arguments{
\item{df}{output from OSCA_par function with permutations}

\item{externalVarName}{string of name of external variable for graph labels; default is blank}

\item{dataName}{string of name of data for graph labels; default is blank}
}
\value{
ggplot object
}
\description{
Function to return a graph comparing percent variance explained in an external phenotype and information lost/percent reduction for both observed and permuted data
}
\references{
Benjamini Y, Hochberg Y. Controlling the false discovery rate: a practical and powerful approach to multiple testing. Journal of the Royal statistical society: series B (Methodological) 57 (1995) 289–300.

Millstein J, Battaglin F, Barrett M, Cao S, Zhang W, Stintzing S, et al. Partition: a surjective mapping approach for dimensionality reduction. Bioinformatics 36 (2019) 676–681. doi:10.1093/bioinformatics/ btz661.
}
\seealso{
OSCA software - \url{https://yanglab.westlake.edu.cn/software/osca}
}
\author{
Katelyn Queen, \email{kjqueen@usc.edu}
}
