% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSCA_singleValue.R
\name{OSCA_singleValue}
\alias{OSCA_singleValue}
\title{OSCA_singleValue}
\usage{
OSCA_singleValue(df, externalVar, oscaPath)
}
\arguments{
\item{df}{n x p dataframe or matrix of gene expression values with no ID column}

\item{externalVar}{vector of length n of external variable values with no ID column}

\item{oscaPath}{absolute path to OSCA software}
}
\value{
Row of OREML output containing percent variance explained in external data and standard error
}
\description{
Function to return the percent variance explained in an external phenotype for a single dataset
}
\references{
Benjamini Y, Hochberg Y. Controlling the false discovery rate: a practical and powerful approach to multiple testing. Journal of the Royal statistical society: series B (Methodological) 57 (1995) 289–300.

Millstein J, Battaglin F, Barrett M, Cao S, Zhang W, Stintzing S, et al. Partition: a surjective mapping approach for dimensionality reduction. Bioinformatics 36 (2019) 676–681. doi:10.1093/bioinformatics/ btz661.
}
\seealso{
OSCA software - \url{https://yanglab.westlake.edu.cn/software/osca}
}
\author{
Katelyn Queen, \email{kjqueen@usc.edu}
}
