% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbes_vd.R
\name{rdbes_vd}
\alias{rdbes_vd}
\title{Table Vessel Details (VD) generation (RDBES process)}
\usage{
rdbes_vd(
  observe_con,
  rdbes_table_cl = NULL,
  rdbes_table_ce = NULL,
  flag,
  hash_algorithms = NULL,
  export_path = NULL
)
}
\arguments{
\item{observe_con}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}} for a connection to the observe database.}

\item{rdbes_table_cl}{{\link[base]{data.frame}} or {\link[tibble]{tibble}} expected. By default NULL. Table CL generated by the function {\link[acdc]{rdbes_cl}}. Warning encrypted vessel code should be according to the database extraction (especially if there was hashing).}

\item{rdbes_table_ce}{{\link[base]{data.frame}} or {\link[tibble]{tibble}} expected. By default NULL. Table CE generated by the function {\link[acdc]{rdbes_ce}}. Warning encrypted vessel code should be according to the database extraction (especially if there was hashing).}

\item{flag}{{\link[base]{integer}} expected. Flag(s) selected associated to the databases queries extractions.}

\item{hash_algorithms}{{\link[base]{integer}} expected. By default NULL. The hashing algorithms to be used for the VDencrypVesIds variable. You can choose any modality of the argument "algo" or the function {\link[digest]{digest}}.}

\item{export_path}{{\link[base]{character}} expected. By default NULL. Directory path associated for the export.}
}
\value{
A R object with the RDBES table VD with potentially a csv extraction associated.
}
\description{
Process for generation and optionally extraction of the RDBES table VD (Vessel Details).
}
\details{
If the arguments "rdbes_table_cl", "rdbes_table_ce" or both are provided, the process generates a table VD according to input data of the data.frame/tibble associated (not from a database SQL extraction).
}
